# apps/docs/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AI Stats Gateway API
  description: >-
    A gateway API for accessing various AI models with OpenAI-compatible
    endpoints.
  version: 1.0.0
  contact:
    name: AI Stats
    url: https://docs.ai-stats.phaseo.app
    email: danielbutler500@gmail.com
servers:
  - url: https://api.phaseo.app/v1
security:
  - BearerAuth: []
paths:
...
  /files:
    post:
      x-beta: true
      tags:
        - Gateway
      operationId: uploadFile
      summary: Upload file
      description: >-
        Compatibility alias for `/batches/files`. Uploads a file for batch
        processing and returns the upstream file metadata. Pass `model` so AI
        Stats can infer the upstream provider. Defaults to OpenAI for legacy
        clients that omit both `model` and `provider`.
      parameters:
        - name: model
          in: query
          required: false
          description: Model id used to infer the upstream batch file provider.
          schema:
            $ref: '#/components/schemas/ModelId'
        - name: provider
          in: query
          required: false
          description: Advanced provider override for the upstream file upload.
          schema:
            type: string
            enum:
              - openai
              - groq
              - together
              - mistral
        - name: x-ai-stats-provider
          in: header
          required: false
          description: Header alternative to the `provider` query parameter.
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          description: Uploaded file metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: listFiles
      summary: List files
      description: >-
        Currently returns `not_supported` on the shared gateway key. Persist
        uploaded file ids and retrieve them directly instead.
      responses:
        '400':
          description: File listing is not supported in the current gateway mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /files/{file_id}:
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: retrieveFile
      summary: Retrieve file
      description: >-
        Retrieves metadata for a file that belongs to the authenticated
        workspace.
      parameters:
        - name: file_id
          in: path
          required: true
          description: The ID of the file to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found or not owned by the workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /files/{file_id}/content:
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: retrieveFileContent
      summary: Retrieve file content
      description: >-
        Retrieves binary content for a previously uploaded file that belongs to
        the authenticated workspace.
      parameters:
        - name: file_id
          in: path
          required: true
          description: The ID of the file to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /batches/files:
    post:
      x-beta: true
      tags:
        - Gateway
      operationId: uploadBatchFile
      summary: Upload batch file
      description: >-
        Uploads a provider file for batch processing and stores ownership in the
        authenticated workspace. Pass `model` so AI Stats can infer the upstream
        provider for provider-backed batch inputs. Defaults to OpenAI for legacy
        clients that omit both `model` and `provider`.
      parameters:
        - name: model
          in: query
          required: false
          description: Model id used to infer the upstream batch file provider.
          schema:
            $ref: '#/components/schemas/ModelId'
        - name: provider
          in: query
          required: false
          description: Advanced provider override for the upstream file upload.
          schema:
            type: string
            enum:
              - openai
              - groq
              - together
              - mistral
        - name: x-ai-stats-provider
          in: header
          required: false
          description: Header alternative to the `provider` query parameter.
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          description: Uploaded file metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /batches/files/{file_id}:
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: retrieveBatchFile
      summary: Retrieve batch file
      description: >-
        Retrieves metadata for a batch file that belongs to the authenticated
        workspace.
      parameters:
        - name: file_id
          in: path
          required: true
          description: The ID of the file to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found or not owned by the workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /batches/files/{file_id}/content:
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: retrieveBatchFileContent
      summary: Retrieve batch file content
      description: >-
        Retrieves binary content for a previously uploaded batch file that
        belongs to the authenticated workspace.
      parameters:
        - name: file_id
          in: path
          required: true
          description: The ID of the file to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /batch/files:
    post:
      x-beta: true
      tags:
        - Gateway
      operationId: uploadBatchFileAlias
      summary: Upload batch file (alias)
      description: Alias of /batches/files.
      parameters:
        - name: model
          in: query
          required: false
          description: Model id used to infer the upstream batch file provider.
          schema:
            $ref: '#/components/schemas/ModelId'
        - name: provider
          in: query
          required: false
          description: Advanced provider override for the upstream file upload.
          schema:
            type: string
        - name: x-ai-stats-provider
          in: header
          required: false
          description: Header alternative to the `provider` query parameter.
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          description: Uploaded file metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /batch/files/{file_id}:
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: retrieveBatchFileAlias
      summary: Retrieve batch file (alias)
      description: Alias of /batches/files/{file_id}.
      parameters:
        - name: file_id
          in: path
          required: true
          description: The ID of the file to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found or not owned by the workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /batch/files/{file_id}/content:
    get:
      x-beta: true
      tags:
        - Gateway
      operationId: retrieveBatchFileContentAlias
      summary: Retrieve batch file content (alias)
      description: Alias of /batches/files/{file_id}/content.
      parameters:
        - name: file_id
          in: path
          required: true
          description: The ID of the file to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    FileUploadRequest:
      type: object
      required:
        - file
        - purpose
      properties:
        file:
          type: string
          format: binary
        purpose:
          type: string
    FileResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        bytes:
          type: integer
        created_at:
          type: integer
        filename:
          type: string
        purpose:
          type: string
        status:
          type: string
        status_details:
          type: object
...
    ModelId:
      type: string
      description: >-
        Model identifier. This is a runtime string so newly released models can
        be used without waiting for an SDK update.
      example: ai21/jamba-large-1.7
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        ok:
          type: boolean
          example: false
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
          example: error_type
        message:
          type: string
          example: Human-readable error message
        description:
          type: string
          example: Additional error details.
        generation_id:
          type: string
          example: G-abc123
        status_code:
          type: integer
          example: 502
        error_type:
          type: string
          enum:
            - user
            - system
          example: system
        error_origin:
          type: string
          enum:
            - user
            - gateway
            - upstream
          example: upstream
        reason:
          type: string
          example: all_candidates_failed
        attempt_count:
          type: integer
          example: 2
        failed_providers:
          type: array
          items:
            type: string
          example:
            - google-ai-studio
            - openai
        failed_statuses:
          type: array
          items:
            type: integer
          example:
            - 403
            - 429
        upstream_error:
          $ref: '#/components/schemas/ErrorUpstreamError'
        failure_sample:
          type: array
          items:
            $ref: '#/components/schemas/ErrorFailureSampleItem'
        provider_failure_diagnostics:
          $ref: '#/components/schemas/ErrorProviderFailureDiagnostics'
        routing_diagnostics:
          $ref: '#/components/schemas/ErrorRoutingDiagnostics'
        provider_candidate_diagnostics:
          $ref: '#/components/schemas/ErrorProviderCandidateDiagnostics'
        provider_enablement:
          $ref: '#/components/schemas/ErrorProviderEnablementDiagnostics'
        missing_pricing_providers:
          type: array
          items:
            type: string
        provider_payment_required_provider:
          type: string
          example: openai
        provider_payment_required_support_notice:
          type: string
          example: >-
            Our upstream provider billing appears to be unavailable. If this
            persists, contact support.
        details:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
    ErrorUpstreamError:
      type: object
      properties:
        code:
          type: string
          nullable: true
          example: PERMISSION_DENIED
        message:
          type: string
          nullable: true
          example: The caller does not have permission.
        description:
          type: string
          nullable: true
        param:
          type: string
          nullable: true
    ErrorFailureSampleItem:
      type: object
      properties:
        provider:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        status:
          type: integer
          nullable: true
        upstream_error_code:
          type: string
          nullable: true
        upstream_error_message:
          type: string
          nullable: true
        upstream_error_description:
          type: string
          nullable: true
        upstream_error_param:
          type: string
          nullable: true
        upstream_payload_preview:
          type: string
          nullable: true
        retryable:
          type: boolean
          nullable: true
      additionalProperties: true
    ErrorProviderFailureDiagnostics:
      type: object
      properties:
        category:
          type: string
          enum:
            - credentials_not_configured
            - credentials_invalid_or_forbidden
            - provider_access_missing
            - region_or_project_restriction
            - model_unavailable_for_endpoint
            - rate_limited
            - server_error
        hint:
          type: string
        provider:
          type: string
          nullable: true
    ErrorRoutingDiagnostics:
      type: object
      properties:
        filterStages:
          type: array
          items:
            type: object
            properties:
              stage:
                type: string
              beforeCount:
                type: integer
              afterCount:
                type: integer
              droppedProviders:
                type: array
                items:
                  type: object
                  properties:
                    providerId:
                      type: string
                      nullable: true
                    reason:
                      type: string
                      nullable: true
                  additionalProperties: true
            additionalProperties: true
      additionalProperties: true
    ErrorProviderCandidateDiagnostics:
      type: object
      properties:
        totalProviders:
          type: integer
        supportsEndpointCount:
          type: integer
        candidateCount:
          type: integer
        droppedUnsupportedEndpoint:
          type: array
          items:
            type: string
        droppedMissingAdapter:
          type: array
          items:
            type: object
            properties:
              providerId:
                type: string
                nullable: true
              endpoint:
                type: string
                nullable: true
            additionalProperties: true
      additionalProperties: true
    ErrorProviderEnablementDiagnostics:
      type: object
      properties:
        capability:
          type: string
        providersBefore:
          type: array
          items:
            type: string
        providersAfter:
          type: array
          items:
            type: string
        dropped:
          type: array
          items:
            type: object
            properties:
              providerId:
                type: string
                nullable: true
              reason:
                type: string
                nullable: true
            additionalProperties: true
      additionalProperties: true
...